Add --force flag to svn export.
Unfortunately, Windows does not, as a general matter, support case-sensitive folders (https://superuser.com/questions/165975/are-all-versions-of-windows-case-insensitive). This caused an unexpected hiccup starting with UE 4.13, which contains both an iOS and an IOS folder in Runtime/Core/. When encountering this issue, svn aborted prematurely, leaving out a significant portion of the Runtime source.

Individual directories can be coerced into case-sensitivity using fsutil, but this does *not* apply recursively. The workaround would be to download one leaf directory at a time, set permissions all up and down the tree, and then repeat. Given how slow GitHub's API response is to these svn requests, this is a non-option.

The best option is adding the --force flag to the svn export command. It overwrites pre-existing folders/files with matching names. In this case, IOS may be overwritten by iOS, resulting in a marginal loss of relevant source data. Still, the savings in complexity over the other solution is worth it.